login / register / save / responsiveness
- login / register functionality.
- save design functionality.
- login / register to appear only after user selects SAVE.
- responsiveness for mobile use.
