+++ src/assets/css/Icon.css
@@ -1,6 +1,10 @@
 .icon {
     width: 40px;
-    height: 40px
+    height: 40px;
+    
+}
+aside .icon {
+    touch-action: none;
 }
 aside .icon:active {
     cursor: grabbing;
@@ -54,3 +58,4 @@ aside .icon:active {
     border: 1px solid lightseagreen;
     cursor: pointer;
 }
+
