+++ src/assets/css/RightMenu.css
@@ -2,10 +2,10 @@
     display: flex;
     flex-direction: column;
     background: #485364;
-    width: 215px;
+    width: 335px;
     height: 100vh;
     padding: 10px 10px 0;
-
+    z-index: 99;
 }
 #rightMenu * {
     box-sizing: border-box;
@@ -53,8 +53,24 @@
     padding-top: 5px;
     padding-bottom: 5px;
     margin-bottom: 5px;
-
+    
 }
+#rightMenu .functions button:not(:first-child) {
+    /*animation: myanim 0.5s 1;*/
+}
+#rightMenu .functions button {
+    animation: myanim 0.5s 1;
+}
+@keyframes myanim {
+    from {
+      opacity: 0;
+      transform: scale(0);
+    }
+    to {
+      opacity: 1;
+      transform: scale(1);
+    }
+  }
 #rightMenu .functions button:hover {
     background-color: #2debd8;
     border-color: #2debd8;
